move arrowcomponent
